Provisioner should now return useful data, yay!
We pass the data when we construct the provisioner FSM and return it
encrypted upon successful authentication.

Still a lot to clean up because I'm not sure combining both roles into
a single set of Phase structs improved clarity.

We also have all the differnt AuthValue's to support.

But at least now we should be able to write a unit test involving two
FSM's: a provisioner and a device.