If necessary, re-transmit the response from the provisionee
Now that the response is available from the fsm, we don't need to
track it in a separate struct. If the incoming PDU's hash is the same
as the previous one, we just respond with the response from the fsm.

One slight change in behavior is that I clear the last_transmit_hash
when the fsm has no response... not sure if that's correct or not.