Add logging of number_of_elements to DeviceInfo
This commit adds info logging of the number_of_elements field for
DeviceInfo.

The motivation for this that even though the number of elements can be
seen when displaying the Bindings of a ProvisionedConfiguration it, at
least for me, makes it a little clearer where these indices are coming
from.