Add comment to crpl field in Composition
This commit adds a comment to the crpl field. This field represents the
minimum number of replay protection list entries on a device.

The motivation for adding this is what it was not clear to me what the
field `crpl` represented. I've looked in the Mesh Profile specification
, and section 4.2.1 "Composition Data" contains a table with this field
but does not explain the abbreviation. I'm only guessing here that it
stands for `Count Replay Protection List`.