Add allow unused_mut attribute for config variable
Currently, when building with the following features a warning is
generated:

cargo b --features="defmt, nrf52840, relay" \
        --features="embassy-time/defmt-timestamp-uptime"

warning: variable does not need to be mutable
  --> src/driver.rs:16:9
   |
16 |     let mut config = nrf_softdevice::Config {
   |         ----^^^^^^
   |         |
   |         help: remove this `mut`
   |
   = note: `#[warn(unused_mut)]` on by default

warning: `btmesh-nrf-softdevice` (lib) generated 1 warning

This commit adds an allow attribute to suppress this warning when these
features are used.

The motivation for this is this warning shows up when building the
firmware for the EclipseCon 2022 Hackaton (when building using local
btmesh chechouts).